+++ fsspec/implementations/libarchive.py
@@ -14,9 +14,10 @@ from ctypes import (
 import libarchive
 import libarchive.ffi as ffi
 
-from fsspec import AbstractFileSystem, open_files
+from fsspec import open_files
+from fsspec.archive import AbstractArchiveFileSystem
 from fsspec.implementations.memory import MemoryFile
-from fsspec.utils import DEFAULT_BLOCK_SIZE, tokenize
+from fsspec.utils import DEFAULT_BLOCK_SIZE
 
 # Libarchive requires seekable files or memory only for certain archive
 # types. However, since we read the directory first to cache the contents
@@ -65,7 +66,7 @@ def custom_reader(file, format_name="all", filter_name="all", block_size=ffi.pag
         yield libarchive.read.ArchiveRead(archive_p)
 
 
-class LibArchiveFileSystem(AbstractFileSystem):
+class LibArchiveFileSystem(AbstractArchiveFileSystem):
     """Compressed archives as a file-system (read-only)
 
     Supports the following formats:
@@ -178,43 +179,6 @@ class LibArchiveFileSystem(AbstractFileSystem):
             }
         )
 
-    def info(self, path, **kwargs):
-        self._get_dirs()
-        path = self._strip_protocol(path)
-        if path in self.dir_cache:
-            return self.dir_cache[path]
-        elif path + "/" in self.dir_cache:
-            return self.dir_cache[path + "/"]
-        else:
-            raise FileNotFoundError(path)
-
-    def ls(self, path, detail=False, **kwargs):
-        self._get_dirs()
-        paths = {}
-
-        for p, f in self.dir_cache.items():
-            p = p.rstrip("/")
-            if "/" in p:
-                root = p.rsplit("/", 1)[0]
-            else:
-                root = ""
-            if root == path.rstrip("/"):
-                paths[p] = f
-            elif all(
-                (a == b)
-                for a, b in zip(path.split("/"), [""] + p.strip("/").split("/"))
-            ):
-                # root directory entry
-                ppath = p.rstrip("/").split("/", 1)[0]
-                if ppath not in paths:
-                    out = {"name": ppath + "/", "size": 0, "type": "directory"}
-                    paths[ppath] = out
-        out = list(paths.values())
-        if detail:
-            return out
-        else:
-            return list(sorted(f["name"] for f in out))
-
     def _open(
         self,
         path,
@@ -239,19 +203,3 @@ class LibArchiveFileSystem(AbstractFileSystem):
                 else:
                     raise ValueError
         return MemoryFile(fs=self, path=path, data=data)
-
-    def ukey(self, path):
-        return tokenize(path, self.fo, self.protocol)
-
-    def _all_dirnames(self, paths):
-        """Returns *all* directory names for each path in paths, including intermediate ones.
-
-        Parameters
-        ----------
-        paths: Iterable of path strings
-        """
-        if len(paths) == 0:
-            return set()
-
-        dirnames = {self._parent(path) for path in paths} - {self.root_marker}
-        return dirnames | self._all_dirnames(dirnames)