+++ docs/source/changelog.rst
@@ -1,9 +1,12 @@
 Changelog
 =========
 
+=======
 Enhancements
 
 - Enable listings cache for HTTP filesystem (#560)
+- Fold ZipFileSystem and LibArchiveFileSystem into a generic implementation and
+  add new TarFileSystem (#561)
 
 
 2021.04.0